engine: switched to integral timing to keep more stable framerates
FossilOrigin-Name: 75e68ed41c6ed2b6e8090adef2ca87d63f0baffdd3a4dfa314f914963d8cf155

+++ source/server/server.h
@@ -81,7 +81,7 @@ extern server_static_t svs;
 // this is GLevel->TicTime value at which we should start rendering
 extern int serverStartRenderFramesTic;
 
-extern double SV_GetFrameTimeConstant () noexcept;
+extern float SV_GetFrameTimeConstant () noexcept;
 
 
 #endif
